test: 프론트 웹소켓 컴포넌트 분리
